Rename EmptyModule into IncompleteModule
